+++ apps/src/templates/sectionProgressV2/TeacherActionsBox.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import i18n from '@cdo/locale';
 import LegendItem, {ITEM_TYPE} from './LegendItem';
-import styles from './section-progress-refresh.module.scss';
+import styles from './progress-table-legend.module.scss';
 import {StrongText} from '@cdo/apps/componentLibrary/typography';
 
 export default function TeacherActionsBox({isViewingLevelProgress}) {
