+++ apps/src/templates/sectionProgressV2/progress-table-legend.module.scss
